+++ Sources/OpenWeatherAgent/OpenWeatherAgent.swift
@@ -2,9 +2,9 @@
 //  OpenWeatherAgent.swift
 //  OpenWeatherAgent
 //
-//  Created by Mikhail Zhigulin in 7531.
+//  Created by Mikhail Zhigulin in 7533.
 //
-//  Copyright © 7531 - 7533 Mikhail A. Zhigulin of Novosibirsk
+//  Copyright © 7533 Mikhail A. Zhigulin of Novosibirsk
 //  Copyright © 7533 PerseusRealDeal
 //
 //  Licensed under the MIT license. See LICENSE file.
@@ -37,16 +37,22 @@ public class OpenWeatherAgent {
 
             guard (200...299).contains(statusCode) else {
                 if statusCode == 404 {
+                    // WRONG: https://api.openweathermap.org/data/999/...
                     throw OpenWeatherAPIClientError.statusCode404
                 }
-                throw OpenWeatherAPIClientError.failedResponse("Status Code: \(statusCode)")
+                // WRONG: https://api.openweathermap.org/...&appid=wrong_api_key
+                let errorDetails = HTTPURLResponse.localizedString(forStatusCode: statusCode)
+                let details = "Status Code: \(statusCode), \(errorDetails)"
+                throw OpenWeatherAPIClientError.failedResponse(details)
             }
 
             return data
 
         } catch let error as URLError {
+            // WRONG: https://apiiiii.openweathermap.org/...
             throw OpenWeatherAPIClientError.failedRequest("URLError: \(error)")
         } catch {
+            // WRONG: something else
             throw error
         }
     }